edit NavBar style
